@@ -52,7 +52,13 @@ public class CheckOutController : Controller
         _logger.LogInformation($"Callback received for id {callbackInfo?.PaymentOrder?.Id}", callbackInfo);
     }
 
-    public async Task<IPaymentOrderResponse> CreateOrUpdatePaymentOrder(bool generatePaymentToken, bool generateRecurrenceToken, bool generateUnscheduledToken,  Uri paymentUrl = null)
+    public async Task<IPaymentOrderResponse> CreateOrUpdatePaymentOrder(bool generatePaymentToken, bool generateRecurrenceToken, bool generateUnscheduledToken)
+    {
+        return await CreateOrUpdatePaymentOrder(generatePaymentToken, generateRecurrenceToken, generateUnscheduledToken, null);
+    }
+
+    public async Task<IPaymentOrderResponse> CreateOrUpdatePaymentOrder(bool generatePaymentToken, bool generateRecurrenceToken, bool generateUnscheduledToken,
+        Uri paymentUrl)
     {
         Uri orderId = null;
         var paymentOrderLink = _cartService.PaymentOrderLink;
@@ -61,10 +67,13 @@ public class CheckOutController : Controller
             orderId = new Uri(paymentOrderLink, UriKind.Relative);
         }
 
-        return orderId == null ? await CreatePaymentOrder(generatePaymentToken, generateRecurrenceToken, generateUnscheduledToken, paymentUrl) : await UpdatePaymentOrder(orderId, generatePaymentToken, generateRecurrenceToken, generateUnscheduledToken);
+        return orderId == null
+            ? await CreatePaymentOrder(generatePaymentToken, generateRecurrenceToken, generateUnscheduledToken, paymentUrl)
+            : await UpdatePaymentOrder(orderId, generatePaymentToken, generateRecurrenceToken, generateUnscheduledToken);
     }
 
-    private async Task<IPaymentOrderResponse> UpdatePaymentOrder(Uri orderId, bool generatePaymentToken,  bool generateRecurrenceToken, bool generateUnscheduledToken, Uri paymentUrl = null)
+    private async Task<IPaymentOrderResponse> UpdatePaymentOrder(Uri orderId, bool generatePaymentToken, bool generateRecurrenceToken, bool generateUnscheduledToken,
+        Uri paymentUrl = null)
     {
         var paymentOrder = await _swedbankPayClient.PaymentOrders.Get(orderId, PaymentOrderExpand.All);
         if (paymentOrder?.Operations.Update == null)
@@ -115,7 +124,13 @@ public class CheckOutController : Controller
         return paymentOrder;
     }
 
-    public async Task<IPaymentOrderResponse> CreatePaymentOrder(bool generatePaymentToken, bool generateRecurrenceToken, bool generateUnscheduledToken, Uri paymentUrl = null)
+    public async Task<IPaymentOrderResponse> CreatePaymentOrder(bool generatePaymentToken, bool generateRecurrenceToken, bool generateUnscheduledToken)
+    {
+        return await CreatePaymentOrder(generatePaymentToken, generateRecurrenceToken, generateUnscheduledToken, null);
+    }
+
+    public async Task<IPaymentOrderResponse> CreatePaymentOrder(bool generatePaymentToken, bool generateRecurrenceToken, bool generateUnscheduledToken,
+        Uri paymentUrl)
     {
         var totalAmount = _cartService.CalculateTotal();
 
@@ -123,7 +138,8 @@ public class CheckOutController : Controller
         var paymentOrderItems = orderItems?.ToList();
         try
         {
-            var paymentOrderRequest = new PaymentOrderRequest(generateRecurrenceToken || generateUnscheduledToken ? Operation.Verify : Operation.Purchase, new Currency("SEK"),
+            var paymentOrderRequest = new PaymentOrderRequest(generateRecurrenceToken || generateUnscheduledToken ? Operation.Verify : Operation.Purchase,
+                new Currency("SEK"),
                 new Amount(totalAmount),
                 new Amount(0), "Test description", "useragent",
                 new Language("sv-SE"),
@@ -202,7 +218,7 @@ public class CheckOutController : Controller
                     CountryCode = new CountryCode("NO")
                 }
             };
-            
+
             var paymentOrder = await _swedbankPayClient.PaymentOrders.Create(paymentOrderRequest, PaymentOrderExpand.All);
 
             _cartService.PaymentOrderLink = paymentOrder?.PaymentOrder.Id.OriginalString;
@@ -218,14 +234,17 @@ public class CheckOutController : Controller
         }
     }
 
-    public async Task<IActionResult> LoadPaymentMenu(bool isRedirect = false, bool generatePaymentToken = false,  bool generateRecurrenceToken = false, bool generateUnscheduledToken = false)
+    
+    public async Task<IActionResult> LoadPaymentMenu(bool isRedirect = false, bool generatePaymentToken = false, bool generateRecurrenceToken = false,
+        bool generateUnscheduledToken = false)
     {
-        var paymentOrder = await CreateOrUpdatePaymentOrder(generatePaymentToken, generateRecurrenceToken, generateUnscheduledToken, _urls.AnonymousCheckoutPaymentUrl);
+        var paymentOrder =
+            await CreateOrUpdatePaymentOrder(generatePaymentToken, generateRecurrenceToken, generateUnscheduledToken, _urls.AnonymousCheckoutPaymentUrl);
         if (isRedirect)
         {
             return Redirect(paymentOrder?.Operations.Redirect?.Href.ToString()!);
         }
-        
+
         var swedbankPaySource = new SwedbankPayCheckoutSource
         {
             JavascriptSource = paymentOrder?.Operations.View?.Href,
